Combined add and edit recipe to one activity, Deleted edit recipe activity to reduce duplication
