+++ Food_Tracker_App/app/src/main/AndroidManifest.xml
@@ -13,11 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.FoodTracker"
         tools:targetApi="31">
-        <activity
-            android:name=".ui.recipes.EditRecipeActivity"
-            android:exported="false"
-            android:label="@string/title_activity_edit_recipe"
-            android:theme="@style/Theme.FoodTracker.NoActionBar" />
+
         <activity
             android:name=".ui.recipes.AddRecipeActivity"
             android:exported="false" />
